selectCurrentTask
